moving transaction-alerts component from /ui/pages/confirm-transaction folder to ui/component/app folder
