+++ ui/components/app/app-components.scss
@@ -103,3 +103,4 @@
 @import 'detected-token/detected-token-selection-popover/index';
 @import 'network-account-balance-header/index';
 @import 'approve-content-card/index';
+@import 'transaction-alerts/transaction-alerts'
