+++ ui/pages/pages.scss
@@ -7,7 +7,6 @@
 @import 'confirm-approve/index';
 @import 'confirm-decrypt-message/confirm-decrypt-message';
 @import 'confirm-encryption-public-key/confirm-encryption-public-key';
-@import 'confirm-transaction-base/transaction-alerts/transaction-alerts';
 @import 'confirmation/confirmation';
 @import 'connected-sites/index';
 @import 'connected-accounts/index';
